+++ lib/src/store.dart
@@ -1088,6 +1088,8 @@ abstract class Vm {
 // /////////////////////////////////////////////////////////////////////////////
 
 abstract class VmEquals<T> {
+  const VmEquals();
+
   bool vmEquals(T other) => identical(this, other);
 }